backend/requirements.txt
flask
flask-cors
langchain
sentence-transformers
faiss-cpu